Fix MaryTTS compatibility layer
MaryTTS receives requests at the `/process` endpoint with URL-encoded parameters either in the URL itself as a GET request, or as a payload with a POST request. Currently, GET requests are processed correctly, but POST requests will be read out as the full URL-encoded string (see #8). This change fixes the processing of the POST request payload to extract the text properly.